Make deep-hist work without accurev diff
The deep-hist algorithm shouldn't need the `accurev diff` command in
order to work. The `pop` and `deep-hist` methods are concerned with
selecting transactions to process while the `diff` method performs
filtering. The `deep-hist` method no longer relies on the success of
`accurev diff` commands which are the basis of the `diff` method.

This should enable the script to work on older accurev versions.