Add deep-hist-only method
The new `deep-hist-only` method is the same as the `deep-hist` method
but doesn't perform any `accurev diff` commands which speed up
transaction selection. This should make this script work with older
versions of Accurev that don't have the `accurev diff` command that can
return a diff between two transactions on the same stream.